# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Data Collection Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/submissions": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "List Submissions",
        "description": "One page of submissions, newest arrival first.\n\n`formId` is the wire form key an op carries, not the form row id, so the\nsame value filters here and identifies a form in a pushed op. An unknown\n`status` is a 422 rather than an empty page — a typo should say so.",
        "operationId": "list_submissions_api_v1_submissions_get",
        "parameters": [
          {
            "name": "formId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Formid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SubmissionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/{submission_id}": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Get Submission",
        "description": "Folded current state plus the raw op log in (counter, deviceId) order.",
        "operationId": "get_submission_api_v1_submissions__submission_id__get",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/{submission_id}/keys": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Get Submission Keys",
        "description": "The wrapped content keys a key holder needs to decrypt this submission.\n\nWrapped copies only, exactly as they were uploaded. Decryption happens in\nthe browser or the desktop app with a private key the server has never held\n(encryption envelope §7) — the console is not a key holder.",
        "operationId": "get_submission_keys_api_v1_submissions__submission_id__keys_get",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionKeysResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/{submission_id}/media": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Get Submission Media",
        "description": "Every file this submission references, and whether each has paired up.\n\nMedia never travels inside the op stream (sync §9): the op carries a\n`mediaId` and the file arrives separately, in either order. So a submission\ncan be complete in every answer and still be waiting for three photographs,\nand this is where that shows. `resolved` is true only when both halves are\nhere — the file is `complete` and an op referencing it has arrived — and\n`pendingCount` is how many are not, so the console can say \"still uploading\"\ninstead of rendering a submission that looks finished and is not.\n\nFor an encrypted file the wrapped media keys come back too, exactly as they\nwere uploaded. The server has never held a private key that opens one, and\nhanding them out is what makes decryption possible at all (envelope §7).",
        "operationId": "get_submission_media_api_v1_submissions__submission_id__media_get",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionMediaResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentKeyView": {
        "properties": {
          "contentKeyId": {
            "type": "string",
            "title": "Contentkeyid"
          },
          "deviceId": {
            "type": "string",
            "title": "Deviceid"
          },
          "wraps": {
            "items": {
              "$ref": "#/components/schemas/WrappedKeyView"
            },
            "type": "array",
            "title": "Wraps"
          }
        },
        "type": "object",
        "required": [
          "contentKeyId",
          "deviceId",
          "wraps"
        ],
        "title": "ContentKeyView",
        "description": "A submission's content key, in the only form the server has it: wrapped."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaKeysView": {
        "properties": {
          "mediaId": {
            "type": "string",
            "title": "Mediaid"
          },
          "contentKeyId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentkeyid"
          },
          "wraps": {
            "items": {
              "$ref": "#/components/schemas/MediaWrappedKeyView"
            },
            "type": "array",
            "title": "Wraps"
          }
        },
        "type": "object",
        "required": [
          "mediaId",
          "contentKeyId",
          "wraps"
        ],
        "title": "MediaKeysView",
        "description": "The wrapped media keys for one file (envelope §6, §7)."
      },
...
      "MediaStatus": {
        "type": "string",
        "enum": [
          "pending",
          "uploading",
          "complete",
          "failed"
        ]
      },
...
      "MediaView": {
        "properties": {
          "mediaId": {
            "type": "string",
            "title": "Mediaid"
          },
          "submissionId": {
            "type": "string",
            "title": "Submissionid"
          },
          "opId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opid"
          },
          "fieldPath": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fieldpath"
          },
          "deviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviceid"
          },
          "mimeType": {
            "type": "string",
            "title": "Mimetype"
          },
          "sizeBytes": {
            "type": "integer",
            "title": "Sizebytes"
          },
          "chunkCount": {
            "type": "integer",
            "title": "Chunkcount"
          },
          "receivedChunks": {
            "type": "integer",
            "title": "Receivedchunks"
          },
          "status": {
            "$ref": "#/components/schemas/MediaStatus"
          },
          "encrypted": {
            "type": "boolean",
            "title": "Encrypted"
          },
          "ciphertextHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ciphertexthash"
          },
          "contentKeyId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentkeyid"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "uploadedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadedat"
          }
        },
        "type": "object",
        "required": [
          "mediaId",
          "submissionId",
          "opId",
          "fieldPath",
          "deviceId",
          "mimeType",
          "sizeBytes",
          "chunkCount",
          "receivedChunks",
          "status",
          "encrypted",
          "ciphertextHash",
          "contentKeyId",
          "resolved",
          "createdAt",
          "uploadedAt"
        ],
        "title": "MediaView",
        "description": "One file as the console sees it."
      },
...
      "MediaWrappedKeyView": {
        "properties": {
          "projectKeyId": {
            "type": "string",
            "title": "Projectkeyid"
          },
          "ephemeralPublic": {
            "type": "string",
            "title": "Ephemeralpublic"
          },
          "nonce": {
            "type": "string",
            "title": "Nonce"
          },
          "wrappedKey": {
            "type": "string",
            "title": "Wrappedkey"
          }
        },
        "type": "object",
        "required": [
          "projectKeyId",
          "ephemeralPublic",
          "nonce",
          "wrappedKey"
        ],
        "title": "MediaWrappedKeyView"
      },
      "MessageError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "MessageError",
        "description": "The body of a refusal that carries prose and nothing to branch on.\n\n`{\"detail\": \"submission not found\"}`. Used where there is exactly one way\nto fail and the status code already says which: a 404 on a read endpoint.\nAnything a client must branch on gets a `reason` field instead — see\n`DeviceRegisterError` and `ProjectKeyError` in `modules/projects/schemas.py`."
      },
      "OpKind": {
        "type": "string",
        "enum": [
          "set",
          "unset",
          "repeat_add",
          "repeat_delete",
          "finalize",
          "reopen"
        ]
      },
...
      "SubmissionDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "projectId": {
            "type": "string",
            "title": "Projectid"
          },
          "formId": {
            "type": "string",
            "title": "Formid"
          },
          "formTitle": {
            "type": "string",
            "title": "Formtitle"
          },
          "formVersion": {
            "type": "integer",
            "title": "Formversion"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          },
          "originDeviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origindeviceid"
          },
          "createdBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdby"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "finalizedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalizedat"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Receivedat"
          },
          "opCount": {
            "type": "integer",
            "title": "Opcount"
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubmissionStateView"
              },
              {
                "type": "null"
              }
            ]
          },
          "ops": {
            "items": {
              "$ref": "#/components/schemas/SubmissionOpView"
            },
            "type": "array",
            "title": "Ops"
          },
          "opsTruncated": {
            "type": "boolean",
            "title": "Opstruncated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "projectId",
          "formId",
          "formTitle",
          "formVersion",
          "status",
          "originDeviceId",
          "createdBy",
          "startedAt",
          "finalizedAt",
          "receivedAt",
          "opCount",
          "state",
          "ops",
          "opsTruncated"
        ],
        "title": "SubmissionDetail"
      },
      "SubmissionKeysResponse": {
        "properties": {
          "submissionId": {
            "type": "string",
            "title": "Submissionid"
          },
          "contentKeys": {
            "items": {
              "$ref": "#/components/schemas/ContentKeyView"
            },
            "type": "array",
            "title": "Contentkeys"
          }
        },
        "type": "object",
        "required": [
          "submissionId",
          "contentKeys"
        ],
        "title": "SubmissionKeysResponse",
        "description": "Every wrapped key needed to decrypt one submission (envelope §7).\n\nA submission built by several devices has one content key per device, all\nwrapped to the same recipients, so a single private key opens every one.\nHanding these out costs nothing: the server has never held the private key\nthat opens them, and neither has whoever is asking, unless they own it."
      },
      "SubmissionListResponse": {
        "properties": {
          "submissions": {
            "items": {
              "$ref": "#/components/schemas/SubmissionSummary"
            },
            "type": "array",
            "title": "Submissions"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "submissions",
          "total",
          "limit",
          "offset"
        ],
        "title": "SubmissionListResponse"
      },
      "SubmissionMediaResponse": {
        "properties": {
          "submissionId": {
            "type": "string",
            "title": "Submissionid"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaView"
            },
            "type": "array",
            "title": "Media"
          },
          "keys": {
            "items": {
              "$ref": "#/components/schemas/MediaKeysView"
            },
            "type": "array",
            "title": "Keys"
          },
          "pendingCount": {
            "type": "integer",
            "title": "Pendingcount"
          }
        },
        "type": "object",
        "required": [
          "submissionId",
          "media",
          "keys",
          "pendingCount"
        ],
        "title": "SubmissionMediaResponse",
        "description": "Every file belonging to one submission, resolved or not."
      },
      "SubmissionOpView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/OpKind"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "value": {
            "title": "Value"
          },
          "encrypted": {
            "type": "boolean",
            "title": "Encrypted"
          },
          "valueCiphertext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valueciphertext"
          },
          "contentKeyId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentkeyid"
          },
          "nonce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonce"
          },
          "deviceId": {
            "type": "string",
            "title": "Deviceid"
          },
          "actorId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actorid"
          },
          "counter": {
            "type": "integer",
            "title": "Counter"
          },
          "wallClock": {
            "type": "string",
            "format": "date-time",
            "title": "Wallclock"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Receivedat"
          },
          "serverSeq": {
            "type": "integer",
            "title": "Serverseq"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "path",
          "value",
          "encrypted",
          "valueCiphertext",
          "contentKeyId",
          "nonce",
          "deviceId",
          "actorId",
          "counter",
          "wallClock",
          "receivedAt",
          "serverSeq"
        ],
        "title": "SubmissionOpView",
        "description": "One row of the raw op log."
      },
      "SubmissionStateView": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "opHighWater": {
            "type": "integer",
            "title": "Ophighwater"
          },
          "computedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Computedat"
          }
        },
        "type": "object",
        "required": [
          "data",
          "opHighWater",
          "computedAt"
        ],
        "title": "SubmissionStateView",
        "description": "The materialised fold: current value per field path."
      },
      "SubmissionStatus": {
        "type": "string",
        "enum": [
          "draft",
          "finalized",
          "in_review",
          "approved",
          "rejected",
          "correction_required"
        ]
      },
      "SubmissionSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "formId": {
            "type": "string",
            "title": "Formid"
          },
          "formTitle": {
            "type": "string",
            "title": "Formtitle"
          },
          "formVersion": {
            "type": "integer",
            "title": "Formversion"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          },
          "originDeviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origindeviceid"
          },
          "opCount": {
            "type": "integer",
            "title": "Opcount"
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Receivedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "formId",
          "formTitle",
          "formVersion",
          "status",
          "originDeviceId",
          "opCount",
          "receivedAt"
        ],
        "title": "SubmissionSummary",
        "description": "One row of the submission list."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WrappedKeyView": {
        "properties": {
          "projectKeyId": {
            "type": "string",
            "title": "Projectkeyid"
          },
          "ephemeralPublic": {
            "type": "string",
            "title": "Ephemeralpublic"
          },
          "nonce": {
            "type": "string",
            "title": "Nonce"
          },
          "wrappedKey": {
            "type": "string",
            "title": "Wrappedkey"
          }
        },
        "type": "object",
        "required": [
          "projectKeyId",
          "ephemeralPublic",
          "nonce",
          "wrappedKey"
        ],
        "title": "WrappedKeyView",
        "description": "One content key wrapped to one recipient project key (envelope §4.3)."
      }
    }
  }
}